# lib/api-zod/src/generated/types/moderationReport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ModerationReport {
  id: string;
  type: string;
  subject: string;
  reason: string;
  reporter: string;
  status: string;
  createdAt: string;
}