# lib/api-zod/src/generated/types/studioArticleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudioArticleInput {
  slug: string;
  title: string;
  excerpt: string;
  /** @nullable */
  content?: string | null;
  category: string;
  authorName: string;
  authorAvatar: string;
  publishedAt: string;
  readTime: number;
  imageUrl: string;
  isBreaking: boolean;
  isTrending: boolean;
}