# lib/api-zod/src/generated/types/studioComment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudioComment {
  id: string;
  articleId: string;
  articleTitle: string;
  user: string;
  /** @nullable */
  avatar: string | null;
  text: string;
  time: string;
  status: string;
}